requirements.txt
streamlit>=1.40.0
anthropic>=0.40.0
elevenlabs>=1.0.0
google-generativeai>=0.8.0
pydub>=0.25.1
Pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
static-ffmpeg>=2.7
